+++ modmain.lua
@@ -70,7 +70,7 @@ modimport("scripts/enchants/wywq")
 modimport("scripts/enchants/lanqiu")
 modimport("scripts/enchants/aiyo")
 modimport("scripts/enchants/fay")
--- modimport("scripts/enchants/yzq")  -- 云中雀已注释
+modimport("scripts/enchants/yzq")
 modimport("scripts/enchants/mgcy")
 modimport("scripts/enchants/kongbai")
 modimport("scripts/enchants/strawberry")